Fixing functions._integrate
The present implementation is not correct. Example

print old_integrate(np.arange(7, dtype=np.float64), 1)
print new_integrate(np.arange(7, dtype=np.float64), 1)

Returns:

[  0.   1.   3.   6.  10.  15.  21.]
[  0.   0.5  2.   4.5  8.  12.5 18.]

The integral of x from 0 to 6 is, in fact, 18, as the new
function returns.

Also, remove the FIXME refering to scipy.integrate.quad, as it
is only for definite integrals of functions, not an indefinite
integral of an array.