BUG: Fix `take` parameter on `upcoef`
`upcoef` function would raise a value error due to a normal division
instead of an integer division. Changing this made it return an empty
array if `take == rec_len - 1`. The cleanest fix was to increment the
right instead of the left boundary, this also allows reconstruction of
the original signal using the sums of upcoef with 'd' and 'a' parts.

NOTE: This means that `upcoef` will return different (but equally
correct) results than before for the same inputs. Doesn't cause any
tests to fail and seems a saner behaviour.