MAINT: Change order in which idwtn processes axes
By changing the order in which `idwtn` processes axes, any iterable used
in the `take` parameter can be used directly instead of first converted
to a tuple and reversed.

Note changes to `test_multidim.py` - these changes shouldn't make a
difference but are necessary to pass the tests. I suspect a bug in
numpy's testing routines.