BUG: Improve input validation for dwtn and idwtn
`idwtn` did not ensure input coefficients were of same size. This
resulted in output size being determined by the order of the dict
constructed, which is unpredictable and could lead to bugs. `idwtn` now
checks to ensure sizes of the accepted input coefficients are identical.

`dwtn` did not enter its main loop if the input was a single python
object. Therefore, the object was returned as-is instead of raising a
ValueError, this is now corrected.