MAINT: Refactor edge conditions in downsampling_convolution
The processing of the edge conditions was not always immediately obvious, I
think the new ones are more clear. The main change is that the filter is
processed in reverse order while 'sliding out', so that input elements are
processed away from the sequence into the extended elements, as they are at the
beginning of the sequence.