DOC: Clean up comments for push to master
Explain stages of input-filter overlap at top of file, as it applies to
multiple functions. Remove comment about padding periodization mode,
to be raised in pull request - why is there padding to multiple of
filter length? Is it mathematically necessary or for MATLAB
compatibility?

Clarify MODE_SYMMETRIC. The extension for ranges larger than twice the
filter length initially confused me.