BLD: Add c99 flag for unix-like compilers
GCC < 5 defaults to c89 as the build language, which doesn't allow variable
declarations in loop headers and the restrict keyword. I don't think any other
compiler has this problem.

We need to subclass build_ext to detect the compiler currently in use (instead
of just the default compiler, which may be overridden).