Added option to skip Matlab compatibility testing
I have added the ability to skip Matlab compatibility testing by
defining the environment variable PYWT_DISABLE_MATLAB. It does not
matter what the variable is set to; if it exists, the test is skipped.
This has been introduced to enable skipping the test even if 'mlab' is
installed since the test takes quite a long time.