MAINT: Be C89 compliant-ish
MSVC 2008 and 2010 must be supported for Python 2.7 and Python 3.4 in the
default windows binary build (including Anaconda).

This requires basically a C89 feature set, plus one-line comments and
`restrict` (as `__restrict`). This also reverts the previous commit, as GNU89
is a superset of MSVCs features.

`restrict` has been `#define`d to something compilers like, and all initial
loop declarations have been removed. Where necessary, extra blocks have been
added to limit variable scope as much as practical. Often this has not been
necessary due to the declaration being in a short loop body. Variable
declarations are moved to the start of blocks.