ENH: Add dwtn based on downcoef_axis
Adds a cython wrapper for downcoef_axis, and a dwtn method in cython. Clocks in
reasonably fast at 6ms on a Haswell i5 for a haar transform of a 256x256 image.
Looks correct at a glance, but hasn't been tested. Could probably be improved
with `cpdef`, `nogil` and other Cython/Numpy optimizations, but I am not
familar with them.

Looks clean to me, and some existing code could probably be replaced by this.