BUG: Fix access to a_info/d_info, temp_output
The info was accessed to calculate an offset, even if no coefficients
were present.

The info should probably be passed in as a pointer instead, given that
the coefs can be NULL, they should also be able to be NULL;

The temporary output was zero'd, instead of copied from memory. This is
the usual use-case, but if there was no temp output the reconstruction
is added to the output. This makes the two consistent. Perhaps the
output should be zero'd instead?