MAINT: Move dwtn out of cython
The number of axes is generally going to be small, so there is little
benefit to moving it to cython. Nothing else is cdef'd, so there is
again little benefit. Profiling shows no difference between them for
2/3D 256^N inputs.

The advantage is better debugging for slightly more of the transform,
and it will be a little more friendly for newcomers to look at and
contribute.