BUG/TST: Fix waverec2 for (None, None, None) level
waverec2 would cause an exception when a level is entirely None. Fixed.

Note there will still be an exception if two consecutive levels are None, and
one of them would be odd-length. In this case, a ValueError would be raised,
because the size of the approximation and detail coefficients will differ by
more than one.