MAINT: Move C-code to sub-folder
I would like to expose more of the external C functions to python for testing
purposes, probably using Cython. The extra testing will be useful when
implementing OpenMP or doing other performance tuning.

Because the .pyx files are converted to .c files in the same directory, this
overwrites the original if they have the same name as the .c file I want to
expose. Moving them to a separate folder alleviates this.