MAINT: Move dwt functions out of _pywt
This is a pure move, it is not expected to compile/run.

Most dwt related functions are split out of _pywt. The public ones are
moved to _dwt.py, and the Cython code moved to _extensions/_dwt.pyx. As
much type-conversion, input checking, etc should happen in the python
modules, and the Cython ones will take MemoryViews and everything in them
will be cdef'd.