MAINT: Unify build process on travis and appveyor
This makes the build process a little more consistent (same packages updated
and installed on both platforms). Also removes the dependency of the appveyor
cache on appveyor.yml, which was preventing me from properly testing it.

Caching on Travis doesn't seem to be working, but we'll see if this helps.