ENH: Benchmark _extensions._dwt, with modes
Benchmarking these functions should hopefully give a closer view of how
performance in the C core changes without requiring a lot of effort to
instrument them separately.

The different modes should have somewhat different performance, especially
for small sizes/large filters. If this doesn't happen, there is too much
overhead to benchmark the convolution directly and this needs to be done
separately.

modes are only set for 1D, since n-dimensional path is not affected by the
mode (it is just passed through to the 1D).