MAINT: Raise a more informative error in the case that DWT detail coefficients are None.
Coefficients as None is allowed by idwtn, but for the multilevel waverecn, the
shape of the detail coefficients arrays are used to make sure the final shape
is correct.  This information is lost if the detail coefficients are None.

The error message in _fix_coeffs is updated to suggest to the user, the
appropriate workaround of using an array of zeros instead.