fix: _check_level too conservative
The current _check_level was using the longest wavelet filter and shortest transform axes
to determine the number of levels.  This assumes the longest wavelet occurs along the
shortest axis which will be overly conservative when this is not the case. Instead,
dwt_max_level should be checked on a per-axis basis.

A test case that would fail under the old behavior was added.