CI, BLD: Run tests on Linux editable installations
This commit performs the following changes:

1. Adds a concurrency key to reduce duplicate runs
2. Moves common environment variables outside of
individual jobs and into the workflow context
for Linux and macOS
3. Adds an extra option to run an editable installation
of PyWavelets under the `--no-build-isolation` flag
with all of the build-time dependencies installed
beforehand, ignoring previous installations as
necessary.

The `pytest --pyargs pywt` invocation is used to
run the test suite against the editable installation
which is similar to the rest of the checks.