FIX: Disable `pytest` caching
The pytest test suite is trying to generate additional
files and it tries to save them to a filesystem that is
for some reason read-only. Since this workflow runs
only in a temporary CI job, the `.pyc` file generation
can be disabled.

Some other fixes are included as well:

1. pydantic<2 not needed anymore
2. Skip the job on forks
3. More readable condition for Anaconda upload
step