MAINT: drop Python 3.11, NumPy 1.25 and Cython 3.0.x support
We need to keep NumPy 1.26.4 for quite a while, but that's the only
1.x version that needs supporting for users that need more time
moving to 2.x

Cython 3.0 is dropped because we want to try moving to the Stable ABI,
and that only has support in 3.1

Python 3.11 is dropped as per the regular schedule (NEP 29 / SPEC 0).

Note that there are no Matplotlib wheels for 32-bit Windows for
Python >=3.12, hence the 32-bit Windows Appveyor job entry is removed.